requirements.txt
telebot==0.0.5
urllib3==2.2.3
watchdog==6.0.0
python-dotenv==1.0.1
Flask==3.0.3
Flask-Session==0.8.0
requests==2.31.0
cryptography==41.0.7
validators==0.22.0
cachetools==5.3.2
ratelimit==2.2.1
schedule==1.2.0